+++ frontend/app/globals.css
@@ -8,6 +8,10 @@
   --background-end-rgb: 255, 255, 255;
 }
 
+.slick-slider{
+  @apply bg-white shadow-sm;
+}
+
 @media (prefers-color-scheme: dark) {
   :root {
     --foreground-rgb: 255, 255, 255;
