+++ frontend/app/layout.js
@@ -6,8 +6,8 @@ import Footer from '@/components/ui/footer'
 const poppins = Poppins({ subsets: ['latin'],weight:['400'] })
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'SnapshortURL',
+  description: 'Streamlined URL Shortening for the Modern Web',
 }
 
 export default function RootLayout({ children }) {
